Enable Superposition
The idea would be to input a color id or code to the scrript, you would have different images of shirts for each color. To make this easier I would import the final color as a layer into your image editing program, design the design onto the tshirt then delete the layer so that your final image is 

1. The same size as the t-shirt
2. Positioned correctly
3. Properly transparent

Now quit bugging me I'll work more this evening.